Fix headless test failure
