+++ tox.ini
@@ -80,7 +80,7 @@ commands_pre =
     headless: pip uninstall -y pytest-qt qtpy pyqt5 pyside2
 commands =
     !headless: pytest {env:PYTEST_PATH:} --color=yes --basetemp={envtmpdir} --cov-report=xml --cov={env:PYTEST_PATH:napari} --ignore tools {posargs}
-    headless: pytest --color=yes --basetemp={envtmpdir} --ignore napari/_vispy --ignore napari/_qt --ignore napari/_tests --ignore tools {posargs}
+    headless: pytest --color=yes --basetemp={envtmpdir} --ignore napari/_vispy --ignore napari/_qt --ignore napari/_tests --ignore-glob napari/plugins/_tests/test_qt*.py --ignore tools {posargs}
 
 
 [testenv:isort]
